Configuration for Tft models

Configuration for Tft models

Usage

tft_config(
  batch_size = 256,
  clip_value = NULL,
  loss = "quantile_loss",
  epochs = 5,
  drop_last = FALSE,
  total_time_steps = NULL,
  num_encoder_steps = NULL,
  quantiles = list(0.5),
  training_tau = 0.3,
  virtual_batch_size = 256,
  valid_split = 0,
  learn_rate = 0.02,
  optimizer = "adam",
  lr_scheduler = NULL,
  lr_decay = 0.1,
  step_size = 30,
  checkpoint_epochs = 10,
  cat_emb_dim = 1,
  hidden_layer_size = 160,
  dropout_rate = 0.3,
  stack_size = 3,
  num_heads = 1,
  verbose = FALSE,
  device = "auto"
)

Arguments

batch_size

(int) Number of examples per batch, large batch sizes are recommended. (default: 1024)

clip_value

If a float is given this will clip the gradient at clip_value. Pass NULL (default) to not clip.

loss

(character or function) Loss function for training within "quantile_loss", "pinball_loss", "rmsse_loss", "smape_loss" (default to quantile_loss)

epochs

(int) Number of training epochs.

drop_last

(bool) Whether to drop last batch if not complete during training

total_time_steps

(int) Size of the look-back time window + forecast horizon in steps. This is the width of Temporal fusion decoder N.

num_encoder_steps

(int) Size of the look-back time window in steps. This is the size of LSTM encoder.

quantiles

(list) list of quantiles forcasts to be used in quantile loss. (default = list(0.5)).

training_tau

(float) training_tau value to be used in pinball loss. (default = 0.3).

virtual_batch_size

(int) Size of the mini batches used for Batch Normalization (default=256)

valid_split

(float) The fraction of the dataset used for validation.

learn_rate

initial learning rate for the optimizer.

optimizer

the optimization method. currently only 'adam' is supported, you can also pass any torch optimizer function.

lr_scheduler

if NULL, (default) no learning rate decay is used. if step decays the learning rate by lr_decay every step_size epochs. It can also be a torch::lr_scheduler function that only takes the optimizer as parameter. The step method is called once per epoch.

lr_decay

multiplies the initial learning rate by lr_decay every step_size epochs. Unused if lr_scheduler is a torch::lr_scheduler or NULL.

step_size

number of epoch before modifying learning rate by lr_decay. Unused if lr_scheduler is a torch::lr_scheduler or NULL.

checkpoint_epochs

checkpoint model weights and architecture every checkpoint_epochs. (default is 10). This may cause large memory usage. Use 0 to disable checkpoints.

cat_emb_dim

(int or list) Embedding size for categorial features, broadcasted to each categorical feature, or per categorical feature when a list of the same size as the categorical features (default=1)

hidden_layer_size

(int)size of the Internal state layer (default=160).

dropout_rate

dropout rate applied to each nn block (default=0.3)

stack_size

(int) Number of self-attention layers to apply (default=3). Use 1 for basic TFT.

num_heads

(int) number of interpretable multi-attention head (default=1)

verbose

(bool) wether to print progress and loss values during training.

device

the device to use for training. cpu or cuda. The default (auto) uses cuda`` if it's available, otherwise uses cpu`.

Value

A named list with all hyperparameters of the TabNet implementation.