Temporal Fusion Transformer model

Fits the Temporal Fusion Transformer for Interpretable Multi-horizon Time Series Forecasting model

Usage

tft_fit(x, ...)

# S3 method for default
tft_fit(x, ...)

# S3 method for recipe
tft_fit(x, df, tft_model = NULL, ..., from_epoch = NULL)

Arguments

x

A recipe specifying a set of preprocessing steps created from recipes::recipe().

The predictor data should be standardized (e.g. centered or scaled). The model treats categorical predictors internally thus, you don't need to make any treatment.

...

Model hyperparameters. See tft_config() for a list of all possible hyperparameters.

df

A data frame containing both the predictors and the outcome.

tft_model

A previously fitted TFT model object to continue the fitting on. if NULL (the default) a brand new model is initialized.

from_epoch

When a tft_model is provided, restore the network weights from a specific epoch. Default is last available checkpoint for restored model, or last epoch for in-memory model.

Value

A TFT model object. It can be used for serialization, predictions, or further fitting.

Fitting a pre-trained model

When providing a parent tft_model parameter, the model fitting resumes from that model weights at the following epoch:

last fitted epoch for a model already in torch context

Last model checkpoint epoch for a model loaded from file

the epoch related to a checkpoint matching or preceding the from_epoch value if provided The model fitting metrics append on top of the parent metrics in the returned TFT model.

Threading

TFT uses torch as its backend for computation and torch uses all available threads by default.

You can control the number of threads used by torch with:

torch::torch_set_num_threads(1)
torch::torch_set_num_interop_threads(1)

Examples

if (torch::torch_is_installed() && FALSE) {
data("vic_elec", package = "tsibbledata")
library(recipes)
rec <- recipe(Demand ~ ., data = vic_elec) %>%
  update_role(Date, new_role="id") %>%
  update_role(Time, new_role="time") %>%
  update_role(Temperature, new_role="observed_input") %>%
  update_role(Holiday, new_role="known_input")
fit <- tft_fit(rec, df = vic_elec, epochs = 1, total_time_steps=10, num_encoder_steps=7)
}